Regionals Build
This is the regionals build of the game.
Added more settings options
Added the title to the title screen
Added the credits menu
Added the questions into the game
Added the background into the game.
Added the option to disable background cars.
Added extra coloring on the results screen.
Fixed a bug where an outofbounds exception would occur at the end of the game.
Removed unused components on the camera.
Added materails that go with the background to the game.